Change name script mariabackup to backup_alldb

+++ backup_alldb.sh
@@ -70,7 +70,7 @@ function check_old_backup(){
     count_full_backup=$(find ${folder_backup} -type f -name ${backup_full_name}-*.gz| wc -l)
     if [ ${count_full_backup} -gt ${full_retention} ]; then
         old_full_number=$(find ${folder_backup} -type f -name ${backup_full_name}-*.gz -printf "%T@ %Tc %p\n" |sort -n |head -n 1 |cut -f4 -d/ |cut -f2 -d- |cut -f1 -d.)
-	if [ -n ${folder_backup} ];then
+	if [ -n ${folder_backup} ] && [ -n ${backup_full_name} ] && [ -n ${old_full_number} ] ;then
 	    rm -rf ${folder_backup}/${backup_full_name}-${old_full_number}*
 	    rm -rf ${folder_backup}/${backup_inc_name}-${old_full_number}*
             echo "Delete full and incremental backup ${old_full_number}" >> log_file
